// apps/admin/src/components/Forms/Region/types.ts
// Enums
export enum LabServiceType {
  CTF = 'CTF',
  LAB = 'LAB',
}

export enum CloudProvider {
  AWS = 'AWS',
  AZURE = 'AZURE',
}

export enum AwsRegion {
  us_east_1 = 'us-east-1',
  us_west_2 = 'us-west-2',
  ap_south_1 = 'ap-south-1',
  ap_southeast_2 = 'ap-southeast-2',
  ap_southeast_1 = 'ap-southeast-1',
  me_central_1 = 'me-central-1',
  il_central_1 = 'il-central-1',
  eu_west_1 = 'eu-west-1',
  sa_east_1 = 'sa-east-1',
  af_south_1 = 'af-south-1',
}

export enum AzureRegion {
  eastus = 'eastus',
  westus2 = 'westus2',
  centralindia = 'centralindia',
  australiacentral = 'australiacentral',
  southeastasia = 'southeastasia',
  uaenorth = 'uaenorth',
  israelcentral = 'israelcentral',
  westeurope = 'westeurope',
  southcentralus = 'southcentralus',
  southafricanorth = 'southafricanorth',
}